Close 17 more vendored conformance gaps (timediff direction, EXISTS rewrite, named windows, RETURNING/trigger ordering)
- timediff(A, B) now renders a modifier that turns B into A with SQLite's
  directional month arithmetic: forward adds clamp overflowed days forward
  (Jan 31 + 1mo = Mar 2) while negative diffs anchor on the later date and
  walk backwards (upstream #8242); sub-day negative intervals keep the sign
  on the whole interval instead of borrowing a day (days truncate toward
  zero, time digits print the absolute remainder). All 12
  timediff-negative-month-roundtrip cases + the 3 scalar-functions-datetime
  negative cases pass.
- EXISTS drops ORDER BY/DISTINCT from the subquery (dropped terms still
  name-resolve; LIMIT/OFFSET stay since OFFSET decides whether a row comes
  out) per select.c's superfluous-clause removal.
- WINDOW definitions whose base is defined nowhere error "no such window";
  a forward reference (base defined later in the same clause) is silently
  ignored, matching SQLite's per-definition resolution order.
- UPDATE ... RETURNING snapshots the row after the write but before the
  AFTER trigger fires; two stale tests updated to the SQLite-cross-checked
  semantics.
- DELETE resets ConflictAlgorithmOverride for its fired triggers (OE_Default
  row-delete coding): OR REPLACE no longer propagates through a DELETE into
  trigger bodies, while explicit inner OR clauses still win.
- TEMP triggers fire before main triggers and, among themselves, in
  creation order (main list stays newest-first).
- WITH-scope CTE names called with arguments reject with "'cte1' is not a
  function" (parser _activeCteNames tracking).

17 expected-failure entries pruned (185 -> 168). Full suite: 7742 passed,
0 failed (net10.0).

+++ src/Ahtola.Tests/ManagedTriggerRowSemanticsTests.cs
@@ -226,10 +226,12 @@ public sealed class ManagedTriggerRowSemanticsTests
         foreach (var sql in setup)
             Execute(managed, sql);
 
+        // SQLite's RETURNING sees the row after the write but before the AFTER
+        // trigger fires (returning.sqltest update-returning-after-trigger cases).
         ReadRows(managed, "UPDATE data SET value = 20 WHERE id = 1 RETURNING id, value")
             .Should()
             .ContainSingle()
-            .Which.Should().Equal(SqlValue.Integer(1), SqlValue.Integer(120));
+            .Which.Should().Equal(SqlValue.Integer(1), SqlValue.Integer(20));
         ReadRows(managed, "SELECT id, value FROM data")
             .Should()
             .ContainSingle()
